Fixed bug in cookie/session handling.
Still have a problem.  If an error occurs inside of server-session, we
don't have access to the original cookie, so we lose our login cookie
if we try to update the profile but make a mistake.

I think a better solution is to figure out how to do the cookie/session
management by adding updates to the end of all the servlets that
(i) creates a session ID and (ii) sets the corresponding cookie
exactly when the session data is nonempty.  The only stumbling block
is that we need a builtin to generate the random number.