Added type system, XSS defense to Weblog
Added type system.  Notice
(infer-types)
(insert-type-checking-and-coersion)
at the bottom of webid.

Types also eliminate XSS.  Right now, XSS is eliminated
as long as the above two lines are run *after* loading a weblog app.
See notes in eliminate-cross-site-scripting function for
details about why this is insufficient; also see the paper.

See the definition of (instrument) to understand all of the security
defenses currently in place with notes about how they interact
and the order in which they should be applied.